Added the "Line quantity" column in the right table.
New features:
- added the new column in the right tableview at the position 2 with name
"Lines quantity";
- some customization of the right tableview.

Also was added source files of the QFileSystemModel, QSortFilterProxyModel
and QStandardItemModel classes to the project dirs.

Need to implement in future:
- the column "Data Modified" in the right tableview now has position 4
(position 3 has in source and parent proxy model), but appropriate data
not filled it. There are need to fix it;
- the selection of the right table was dropped when increase its column
quntity by implementing the RightViewProxyModel::columnCount() method. There
are need to fix it;
- maybe delete the thin lines in the right tableview;
- maybe add the checkboxes for every item in the right tableview.

Signed-off-by: Oleg <olgerd27@gmail.com>